+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+## 10.20.1 (01-07-2026)
+
+### Bugfix: 2
+- [#40934](https://parermine.regione.emilia-romagna.it/issues/40934) Correzione apertura pagine ricerca dati specifici
+- [#40927](https://parermine.regione.emilia-romagna.it/issues/40927) Modifica ai controlli al recupero di file da Sacer
+
 ## 10.20.0 (26-06-2026)
 
 ### Bugfix: 5